+++ lading/src/generator/procfs.rs
@@ -7,12 +7,13 @@
 //! Data types for fields generally assume a 64-bit architecture, rather than
 //! use Rust's C-compatible data types.
 
-use ::std::{num::NonZeroU32, path::PathBuf};
+use std::{fs::File, io::Write, num::NonZeroU32, path::PathBuf};
 
-use ::rand::SeedableRng;
-use ::serde::Deserialize;
+use lading_payload::procfs;
+use rand::rngs::StdRng;
+use rand::SeedableRng;
+use serde::Deserialize;
 
-use super::General;
 use crate::signals::Shutdown;
 
 #[derive(::thiserror::Error, Debug)]
@@ -30,8 +31,8 @@ pub struct Config {
     pub seed: [u8; 32],
     /// Root path for `procfs` filesystem
     pub root: PathBuf,
-    /// Upper bound on processes created
-    pub max_processes: NonZeroU32,
+    /// Total number of processes created
+    pub total_processes: NonZeroU32,
 }
 
 /// The procfs generator.
@@ -39,36 +40,84 @@ pub struct Config {
 /// Generates a fake `procfs` filesystem. Currently incomplete.
 #[allow(dead_code)]
 #[derive(Debug)]
-pub struct Procfs {
-    config: Config,
+pub struct ProcFs {
     shutdown: Shutdown,
 }
 
-impl Procfs {
+impl ProcFs {
     /// Create a new [`Procfs`] generator.
     ///
     /// # Errors
     ///
     /// Returns an error if the config cannot be serialized.
-    pub fn new(general: General, config: Config, shutdown: Shutdown) -> Result<Self, Error> {
-        // TODO(geoffrey.oxberry@datadoghq.com): A more feature-ful version of a
-        // procfs generator should check whether the maximum number of arguments
-        // setting doesn't violate `sysconf` settings, e.g., `ARG_MAX`.
+    pub fn new(config: &Config, shutdown: Shutdown) -> Result<Self, Error> {
+        let mut rng = StdRng::from_seed(config.seed);
 
-        // TODO(geoffrey.oxberry@datadoghq.com): Use this concrete RNG to build
-        // procfs payloads that are parametrized by an abstract type satisfying
-        // the trait bound `rand::Rng + ?Sized`.
-        let mut _rng = ::rand::rngs::StdRng::from_seed(config.seed);
+        let total_processes = config.total_processes.get();
+        let mut processes = procfs::fixed(&mut rng, total_processes as usize);
 
-        let mut labels = vec![
-            ("component".to_string(), "generator".to_string()),
-            ("component_name".to_string(), "procfs".to_string()),
-        ];
-        if let Some(id) = general.id {
-            labels.push(("id".to_string(), id));
+        for process in processes.drain(..) {
+            let pid = process.pid;
+
+            let mut pid_root: PathBuf = config.root.clone();
+            pid_root.push(format!("{pid}"));
+
+            // /proc/{pid}/cmdline
+            {
+                let mut path: PathBuf = pid_root.clone();
+                path.push("cmdline");
+
+                let mut file = File::create(path)?;
+                file.write_all(process.cmdline.as_bytes())?;
+            }
+
+            // /proc/{pid}/comm
+            {
+                let mut path: PathBuf = pid_root.clone();
+                path.push("comm");
+
+                let mut file = File::create(path)?;
+                file.write_all(process.comm.as_bytes())?;
+            }
+
+            // /proc/{pid}/io
+            {
+                let mut path: PathBuf = pid_root.clone();
+                path.push("io");
+
+                let mut file = File::create(path)?;
+                file.write_all(format!("{}", process.io).as_bytes())?;
+            }
+
+            // /proc/{pid}/stat
+            {
+                let mut path: PathBuf = pid_root.clone();
+                path.push("stat");
+
+                let mut file = File::create(path)?;
+                file.write_all(format!("{}", process.stat).as_bytes())?;
+            }
+
+            // /proc/{pid}/statm
+            {
+                let mut path: PathBuf = pid_root.clone();
+                path.push("statm");
+
+                let mut file = File::create(path)?;
+                file.write_all(format!("{}", process.statm).as_bytes())?;
+            }
+
+            // /proc/{pid}/status
+            {
+                let mut path: PathBuf = pid_root.clone();
+                path.push("status");
+
+                let mut file = File::create(path)?;
+                file.write_all(format!("{}", process.status).as_bytes())?;
+            }
         }
 
-        Ok(Self { config, shutdown })
+        Ok(Self { shutdown })
     }
 
     /// Run [`Procfs`] generator to completion or until a shutdown signal is
